Add files via upload
I have used HTML and CSS to create a form where you can Add, edit, and remove the details of an employee. On deletion of an employee's details a conformation message will pop up asking that wether you are sure about deleting the details of that employee. Form validation is implemented using javascript. For storing an employee's details it is mandatory to fill the Full name of the employee.